requirements/development.txt
# Development requirements

-r base.txt

django-debug-toolbar==3.7.0
pytest-django==4.5.2
mixer==7.2.2
pytest-cov==4.0.0
flake8==6.0.0
flake8-isort==5.0.3
black==22.10.0
pylint-django==2.5.3
pre-commit==2.20.0
factory-boy==3.2.1
